Website updates from Charles.  All complete except cutout service step & filter explanation and link to youtube from spinning cube.
Here is his list:

* On home page (and all others), can the cubes on left be ³centered
vertically² with the text paragraph so it is not squeezed on top?

* On http://test.turbulence.pha.jhu.edu/instructionswebserv.aspx
There is a big cluster sketch at bottom. It can be deleted..

* On http://test.turbulence.pha.jhu.edu/cutout.aspx
The new ³step² and ³filter² boxes are great. I wonder if we need to add
1-2
Explanatory sentences on this page on how this works, what sort of filter
is done,
Etc.
Also, right now I suppose the ³step² only works for all spatial
directions. The way it is presented, it seems as if it also steps through
time with same step. I think we need
steps for time separately from space. Actually, the cleanest would be to
have a step for
All 4 categories, t x y z . And for filter, indicate ³Spatial 3D box
filter size²
where we only allow same filter size in all 3 spatial directions. We may
wish to discuss this at next meeting ..

* A question: I think now the ³email us for an authorization token² is
rather lost somewhere. I could not even find it. I think we should add
this to
http://test.turbulence.pha.jhu.edu/instructionswebserv.aspx
Somewhere visible so it makes sense. Maybe right after
³..but one should be able to get it working with any language supporting
Web Services.²
Add:
³Single requests for data at more than 4096 points require an
authorization token²
And make a link to the page we have for this:
http://test.turbulence.pha.jhu.edu/help/authtoken.aspx
(where also the right picture is too squeezed and has the cluster..)

*http://test.turbulence.pha.jhu.edu/matlabanalysistools.aspx
On left cubes are squeezed in strange shape and cluster sketch can be
deleted

*On
http://test.turbulence.pha.jhu.edu/help/c-fortran/
I wonder if people easily discern that they are supposed to click on the
download to get the tar. Or zip for C and Fortran folders. Could we make
the first line larger and more explicit, something like ³Download folder
here² or something invitingŠ
(same in the Matlab
http://test.turbulence.pha.jhu.edu/help/matlab/

* On documentation
http://test.turbulence.pha.jhu.edu/datasets.aspx
I think we can make fonts of text smaller, like the default in other
pages..

* In publications
http://test.turbulence.pha.jhu.edu/publications.aspx
A lot of the links open in same browser window. I would say let¹s make all
of the papers and links ope in new window.

* In http://test.turbulence.pha.jhu.edu/analysisdoc.aspx
let¹s place the
"For detailed formulae used in the functions, DOWNLOAD DOCUMENT ³
Above the table below the title.
Also, squares on left are squeezed..
And, on bottom explanations, make fonts smaller

* In http://test.turbulence.pha.jhu.edu/citing.aspx
Make fonts uniform, a bit smaller.
And,
Add
"For articles that use the mhd flow data, we suggest also including:"
..Authors and link to PDF README as in the channel

"For articles that use the mixing flow data, we suggest also including:"
..Authors and link to PDF README as in the channel

* Dropdown for links:
The Center for Turbulence R and the NASA sites should be listed under
³external sites²

* Under visualizations, the rotating cube is really very slow to load.
Somehow it would be nice to have it the same as the other 2. Could we
simply upload onto youtube or something like that?

* Under people and credits,
let¹s place ³External collaborators²
Just below ³postdocs² and add Shiyi Chen and Ethan Vishniac there,
eliminate the
³associated faculty² category

* under postdocs, please add ³Jin Lee²

* ³Others² should be changed to ³Technical staff² or something like that.